+++ internal/selfupdate/selfupdate_suite_test.go
@@ -1,0 +1,13 @@
+package selfupdate_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestSelfupdate(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Selfupdate Suite")
+}